Slight refactor of application control
