+++ psi/application/__init__.py
@@ -358,63 +358,6 @@ def load_paradigm_descriptions():
         importlib.import_module(description)
 
 
-def add_default_options(parser):
-    import argparse
-
-    class CalibrationAction(argparse.Action):
-        def __call__(self, parser, namespace, value, option_string=None):
-            path = Path(value)
-            if not path.exists():
-                path = namespace.io / path
-                path = path.with_suffix('.json')
-                if not path.exists():
-                    raise ValueError('%s does not exist'.format(value))
-            setattr(namespace, self.dest, value)
-
-    try:
-        default_io = get_default_io()
-    except ValueError:
-        default_io = None
-    parser.add_argument('pathname', type=str, help='Filename', nargs='?')
-    parser.add_argument('--io', type=str, default=default_io,
-                        help='Hardware configuration')
-    parser.add_argument('--calibration', type=str, help='Hardware calibration',
-                        action=CalibrationAction)
-    parser.add_argument('--debug', default=True, action='store_true',
-                        help='Debug mode?')
-    parser.add_argument('--debug-warning', default=False, action='store_true',
-                        help='Show warnings?')
-    parser.add_argument('--debug-level-console', type=str, default='INFO',
-                        help='Logging level for console')
-    parser.add_argument('--debug-level-file', type=str, default='INFO',
-                        help='Logging level for file')
-    parser.add_argument('--debug-exclude', type=str, nargs='*',
-                        help='Names to exclude from debugging')
-    parser.add_argument('--pdb', default=False, action='store_true',
-                        help='Autolaunch PDB?')
-    parser.add_argument('--no-preferences', default=False, action='store_true',
-                        help="Don't load existing preference files")
-    parser.add_argument('--no-layout', default=False, action='store_true',
-                        help="Don't load existing layout files")
-    parser.add_argument('-c', '--commands', nargs='+', default=[],
-                        help='Commands to invoke')
-    parser.add_argument('-p', '--preferences', type=str, nargs='?',
-                        help='Preferences file')
-    parser.add_argument('-l', '--layout', type=str, nargs='?',
-                        help='Layout file')
-    parser.add_argument('--profile', action='store_true', help='Profile app')
-
-
-def parse_args(parser):
-    args = parser.parse_args()
-    if args.calibration is None:
-        try:
-            args.calibration = get_default_calibration(args.io)
-        except ValueError as e:
-            log.warn(str(e))
-    return args
-
-
 def list_io_templates():
     io_template_path = Path(__file__).parent.parent / 'templates' / 'io'
     return list(io_template_path.glob('*.enaml'))
